corrected bug
You just forgot to include the file "sensor.h". 
In this case, you need to create a class that depends on another one (RangeSensor is a derived class while Sensor is a base class). 
In this process, you have to give to your compiler the object you are using, you need to let it know what "Sensor" is. To do so, you need to include the .h file containing the class Sensor