Problem solved
The problem here was that you were declaring the object RangeSensor without creating a variable for it. You call the constructor and it worked fine but no variable was created. 

Once you started the code. in the robot you have a vector of pointers to sensors. However, since you didn't create any variable then the pointer is not valid.